config Ecu for project

+++ Config/Port_Cfg.h
@@ -21,7 +21,7 @@ extern "C" {
  *	- High nible: index port (A=0, B=1 , C=2 ,...)
  *	- Low nible: pin number  (0....15)
  * --------------------------------------------------------- */
-#define PORT_ENCODE_PIN(port_letter, pin_num)		((()((port_letter) - 'A') << 4 ) | (()(pin_num) & 0x0Fu))
+#define PORT_ENCODE_PIN(port_letter, pin_num)		(((uint8)((port_letter) - 'A') << 4 ) | ((uint8)(pin_num) & 0x0Fu))
 #define PORTA_PIN(n)		PORT_ENCODE_PIN('A', (n))
 #define PORTB_PIN(n)		PORT_ENCODE_PIN('B', (n))
 #define PORTC_PIN(n)		PORT_ENCODE_PIN('C', (n))
